+++ src/vs/sessions/sessions.web.main.ts
@@ -165,13 +165,13 @@ import './contrib/remoteAgentHost/browser/hostFilter.contribution.js';
 // build. The desktop mode + model pickers are gated off on phone via
 // `when: IsPhoneLayoutContext.negate()`, so the two registrations are
 // mutually exclusive at the action-menu level.
-import './contrib/chat/browser/agentHost/mobileChatInputConfigPicker.js';
+import './contrib/chat/browser/mobile/mobileChatInputConfigPicker.js';
 
 // Phone-only presenter for the workbench `ChatInputPart`'s Mode + Model
 // pickers. Replaces the desktop popups with the same bottom-sheet
 // experience used by the empty new-chat input, applied to the
 // already-opened chat input. Web-only for the same reason as above.
-import './contrib/chat/browser/mobileChatPhoneInputPresenter.js';
+import './contrib/chat/browser/mobile/mobileChatPhoneInputPresenter.js';
 
 // Mobile-aware Copilot permission picker. Replaces the desktop
 // permission picker registration (which the shared contribution
